+++ docs/changes.md
@@ -1,5 +1,8 @@
 # 最新动态
 
+2026/8/26
+  * fix children_layouter_default hflow vflow(感谢兆坤提供补丁)
+
 2026/8/25
   * 修正edit失去焦点后没有重置input_method中的widget指针的问题(感谢福明提供补丁)
   * 修复clang-cl下SDL编译错误(感谢均权提供补丁)